Created the Song Loader
This is the initlal song loader using Spotipy. It currently loads in the track IDs that are found in a specified playlist. NB: Authentication is required. This is at a very early stage.

TODO: Scrape the ids, allow login via text input. Also, put the whole thing in a class. Create a method to allow creation of playlists too.