Added the class structure to the code
By creating an instance of the Spotify Loader class, the get_playlist_track_list method can be called to list all of the track IDs of the songs in the playlist. This class just obtains the track IDs from various methods, and it will also later be used to interface with Spotify easily.