Bad genetic algorithm first commit
This is a genetic algorithm to reorder tracks and find the best possible solution to make the music flow. Some of this code needs refactoring, and other parts need to be totally replaced. It is a proof of concept though. 

TODO: Interface with spotify to play a playlist. Once the playlist playing happens, we will be able to test it using the music, rather than a few IDs.