Upgrades to natural selection/representation
Lots done here!

Firstly, the population is now sorted in order of fitness - the mix with the lowest value for the fitness function is stored at the front of the list, and the whole list is in ascending order of fitness function values. The best solution can therefore be found by finding the first item in this list. When adding new children to the list, they are added in the correct location by finding the location that they should be inserted at (using linear search, so this might be able to be improved with some clever binary search type algorithm, but maybe not).

Secondly, natural selection now works by selecting two random parents from the population (that are in the top levels of fitness). This adds a bit more randomness to the whole system.

TODO: Population culling - occasionally killing off the bad mixes; more children produced per generation (call the breed function a random number of times per generation).